+++ lib/teiserver/communication/queries/room_queries.ex
@@ -34,38 +34,41 @@ defmodule Teiserver.Communication.RoomQueries do
   def _where(query, _, "Any"), do: query
 
   def _where(query, :id, id) do
-    from rooms in query,
+    from(rooms in query,
       where: rooms.id == ^id
+    )
   end
 
   def _where(query, :id_in, id_list) do
-    from rooms in query,
+    from(rooms in query,
       where: rooms.id in ^id_list
+    )
   end
 
   def _where(query, :name, name) do
-    from rooms in query,
+    from(rooms in query,
       where: rooms.name == ^name
+    )
   end
 
-
   def _where(query, :inserted_after, timestamp) do
-    from rooms in query,
+    from(rooms in query,
       where: rooms.inserted_at >= ^timestamp
+    )
   end
 
   def _where(query, :inserted_before, timestamp) do
-    from rooms in query,
+    from(rooms in query,
       where: rooms.inserted_at < ^timestamp
+    )
   end
-
 
   @spec do_order_by(Ecto.Query.t(), list | nil) :: Ecto.Query.t()
   defp do_order_by(query, nil), do: query
 
   defp do_order_by(query, params) when is_list(params) do
     params
-    |> List.wrap
+    |> List.wrap()
     |> Enum.reduce(query, fn key, query_acc ->
       _order_by(query_acc, key)
     end)
@@ -73,25 +76,28 @@ defmodule Teiserver.Communication.RoomQueries do
 
   @spec _order_by(Ecto.Query.t(), any()) :: Ecto.Query.t()
   def _order_by(query, "Name (A-Z)") do
-    from rooms in query,
+    from(rooms in query,
       order_by: [asc: rooms.name]
+    )
   end
 
   def _order_by(query, "Name (Z-A)") do
-    from rooms in query,
+    from(rooms in query,
       order_by: [desc: rooms.name]
+    )
   end
 
   def _order_by(query, "Newest first") do
-    from rooms in query,
+    from(rooms in query,
       order_by: [desc: rooms.inserted_at]
+    )
   end
 
   def _order_by(query, "Oldest first") do
-    from rooms in query,
+    from(rooms in query,
       order_by: [asc: rooms.inserted_at]
+    )
   end
-
 
   @spec do_preload(Ecto.Query.t(), List.t() | nil) :: Ecto.Query.t()
   defp do_preload(query, nil), do: query