+++ lib/teiserver/contexts/communication.ex
@@ -1,7 +1,8 @@
 defmodule Teiserver.Communication do
   @moduledoc """
   The contextual module for:
-  - `Teiserver.Communication.ChatMessage`
+  - `Teiserver.Communication.Room`
+  - `Teiserver.Communication.RoomMessage`
   - `Teiserver.Communication.DirectMessage`
   - `Teiserver.Communication.LobbyMessage`
   - `Teiserver.Communication.PartyMessage`
@@ -14,37 +15,44 @@ defmodule Teiserver.Communication do
   defdelegate room_query(args), to: RoomQueries
 
   @doc section: :room
-  @spec list_rooms() :: [Room.t]
+  @spec list_rooms() :: [Room.t()]
   defdelegate list_rooms(), to: RoomLib
 
   @doc section: :room
-  @spec list_rooms(list) :: [Room.t]
+  @spec list_rooms(list) :: [Room.t()]
   defdelegate list_rooms(args), to: RoomLib
 
   @doc section: :room
-  @spec get_room!(non_neg_integer(), list) :: Room.t
+  @spec get_room!(non_neg_integer(), list) :: Room.t()
   defdelegate get_room!(room_id, query_args \\ []), to: RoomLib
 
   @doc section: :room
-  @spec get_room(non_neg_integer(), list) :: Room.t | nil
+  @spec get_room(non_neg_integer(), list) :: Room.t() | nil
   defdelegate get_room(room_id, query_args \\ []), to: RoomLib
 
   @doc section: :room
-  @spec create_room(map) :: {:ok, Room.t} | {:error, Ecto.Changeset}
+  @spec get_room_by_name(String.t()) :: Room.t() | nil
+  defdelegate get_room_by_name(room_name), to: RoomLib
+
+  @doc section: :room
+  @spec create_room(map) :: {:ok, Room.t()} | {:error, Ecto.Changeset}
   defdelegate create_room(attrs \\ %{}), to: RoomLib
 
   @doc section: :room
-  @spec update_room(Room, map) :: {:ok, Room.t} | {:error, Ecto.Changeset}
+  @spec get_or_create_room(String.t()) :: Room.t()
+  defdelegate get_or_create_room(room_name), to: RoomLib
+
+  @doc section: :room
+  @spec update_room(Room, map) :: {:ok, Room.t()} | {:error, Ecto.Changeset}
   defdelegate update_room(room, attrs), to: RoomLib
 
   @doc section: :room
-  @spec delete_room(Room.t) :: {:ok, Room.t} | {:error, Ecto.Changeset}
+  @spec delete_room(Room.t()) :: {:ok, Room.t()} | {:error, Ecto.Changeset}
   defdelegate delete_room(room), to: RoomLib
 
   @doc section: :room
-  @spec change_room(Room.t, map) :: Ecto.Changeset
+  @spec change_room(Room.t(), map) :: Ecto.Changeset
   defdelegate change_room(room, attrs \\ %{}), to: RoomLib
-
 
   alias Teiserver.Communication.{RoomMessage, RoomMessageLib, RoomMessageQueries}
 
@@ -53,37 +61,36 @@ defmodule Teiserver.Communication do
   defdelegate room_message_query(args), to: RoomMessageQueries
 
   @doc section: :room_message
-  @spec list_room_messages() :: [RoomMessage.t]
+  @spec list_room_messages() :: [RoomMessage.t()]
   defdelegate list_room_messages(), to: RoomMessageLib
 
   @doc section: :room_message
-  @spec list_room_messages(list) :: [RoomMessage.t]
+  @spec list_room_messages(list) :: [RoomMessage.t()]
   defdelegate list_room_messages(args), to: RoomMessageLib
 
   @doc section: :room_message
-  @spec get_room_message!(non_neg_integer(), list) :: RoomMessage.t
+  @spec get_room_message!(non_neg_integer(), list) :: RoomMessage.t()
   defdelegate get_room_message!(room_message_id, query_args \\ []), to: RoomMessageLib
 
   @doc section: :room_message
-  @spec get_room_message(non_neg_integer(), list) :: RoomMessage.t | nil
+  @spec get_room_message(non_neg_integer(), list) :: RoomMessage.t() | nil
   defdelegate get_room_message(room_message_id, query_args \\ []), to: RoomMessageLib
 
   @doc section: :room_message
-  @spec create_room_message(map) :: {:ok, RoomMessage.t} | {:error, Ecto.Changeset}
+  @spec create_room_message(map) :: {:ok, RoomMessage.t()} | {:error, Ecto.Changeset}
   defdelegate create_room_message(attrs \\ %{}), to: RoomMessageLib
 
   @doc section: :room_message
-  @spec update_room_message(RoomMessage, map) :: {:ok, RoomMessage.t} | {:error, Ecto.Changeset}
+  @spec update_room_message(RoomMessage, map) :: {:ok, RoomMessage.t()} | {:error, Ecto.Changeset}
   defdelegate update_room_message(room_message, attrs), to: RoomMessageLib
 
   @doc section: :room_message
-  @spec delete_room_message(RoomMessage.t) :: {:ok, RoomMessage.t} | {:error, Ecto.Changeset}
+  @spec delete_room_message(RoomMessage.t()) :: {:ok, RoomMessage.t()} | {:error, Ecto.Changeset}
   defdelegate delete_room_message(room_message), to: RoomMessageLib
 
   @doc section: :room_message
-  @spec change_room_message(RoomMessage.t, map) :: Ecto.Changeset
+  @spec change_room_message(RoomMessage.t(), map) :: Ecto.Changeset
   defdelegate change_room_message(room_message, attrs \\ %{}), to: RoomMessageLib
-
 
   alias Teiserver.Communication.{DirectMessage, DirectMessageLib, DirectMessageQueries}
 
@@ -92,40 +99,39 @@ defmodule Teiserver.Communication do
   defdelegate direct_message_query(args), to: DirectMessageQueries
 
   @doc section: :direct_message
-  @spec list_direct_messages() :: [DirectMessage.t]
+  @spec list_direct_messages() :: [DirectMessage.t()]
   defdelegate list_direct_messages(), to: DirectMessageLib
 
   @doc section: :direct_message
-  @spec list_direct_messages(list) :: [DirectMessage.t]
+  @spec list_direct_messages(list) :: [DirectMessage.t()]
   defdelegate list_direct_messages(args), to: DirectMessageLib
 
   @doc section: :direct_message
-  @spec get_direct_message!(non_neg_integer(), list) :: DirectMessage.t
+  @spec get_direct_message!(non_neg_integer(), list) :: DirectMessage.t()
   defdelegate get_direct_message!(direct_message_id, query_args \\ []), to: DirectMessageLib
 
   @doc section: :direct_message
-  @spec get_direct_message(non_neg_integer(), list) :: DirectMessage.t | nil
+  @spec get_direct_message(non_neg_integer(), list) :: DirectMessage.t() | nil
   defdelegate get_direct_message(direct_message_id, query_args \\ []), to: DirectMessageLib
 
   @doc section: :direct_message
-  @spec create_direct_message(map) :: {:ok, DirectMessage.t} | {:error, Ecto.Changeset}
+  @spec create_direct_message(map) :: {:ok, DirectMessage.t()} | {:error, Ecto.Changeset}
   defdelegate create_direct_message(attrs \\ %{}), to: DirectMessageLib
 
   @doc section: :direct_message
-  @spec update_direct_message(DirectMessage, map) :: {:ok, DirectMessage.t} | {:error, Ecto.Changeset}
+  @spec update_direct_message(DirectMessage, map) ::
+          {:ok, DirectMessage.t()} | {:error, Ecto.Changeset}
   defdelegate update_direct_message(direct_message, attrs), to: DirectMessageLib
 
   @doc section: :direct_message
-  @spec delete_direct_message(DirectMessage.t) :: {:ok, DirectMessage.t} | {:error, Ecto.Changeset}
+  @spec delete_direct_message(DirectMessage.t()) ::
+          {:ok, DirectMessage.t()} | {:error, Ecto.Changeset}
   defdelegate delete_direct_message(direct_message), to: DirectMessageLib
 
   @doc section: :direct_message
-  @spec change_direct_message(DirectMessage.t, map) :: Ecto.Changeset
+  @spec change_direct_message(DirectMessage.t(), map) :: Ecto.Changeset
   defdelegate change_direct_message(direct_message, attrs \\ %{}), to: DirectMessageLib
-
-
 
   # Lobby chat
   # Party chat
-
 end